feat: sweeping changes, ci files, chart index
This is a major change to the chart. It includes additions to several areas as well as name changes. This adds some basic files required in CI which will likely change soon, as well as the initial chart repo index.yaml. We lastly renamed the chart from experimental to authelia as it's close to becoming a beta.