feat(authelia): add hashicorp vault injector
This commit adds the HashiCorp Vault injector configuration. Additionally it refactors the ConfigMap and Secret generation methods. It introduces several major changes to the values.yaml which should be considered breaking. Several providers now have an 'enabled' option and some have 'enabledSecret'. How they operate is 'enabled' renders the values of the section in the ConfigMap, and the enabledSecret option forces the secret to be used - this is used for redis and SMTP. Lastly it removes the commented sections of a majority of the commented configuration in values.yaml making it more compatible with various GitOps solutions.

Fixes #12