add enabled value to schema
This adds a boolean value `enabled` to the chart schema
which can be used when this chart is pulled in as a dependency.

See https://helm.sh/docs/chart_best_practices/dependencies/#conditions-and-tags
for details.

This is required since additionalProperties is false, so downstream charts cannot
define `authelia.enabled` without this.

Signed-off-by: Jordan Evans <jevans@linuxfoundation.org>