feat(authelia): Additional objects
This adds a template to create additional objects that are templated
with the charts values. This is useful for creating custom resources
that are not included in the chart by default.

Allowed are either yaml objects or string objects. The string objects
are evaluated as templates using `tpl`.

```yaml
extraObjects:
  - apiVersion: v1
    kind: ConfigMap
    metadata:
      name: custom-config
    ...
  - |
    apiVersion: v1
    kind: Secret
    metadata:
      name: custom-secret-{{ .Release.Name }}
    ...
```